Joren/dev steady state scheduler rebased (#82)
* update layer stacks for resnet example

* initial steady state scheduler without memory and port allocation optimization

* remove 11 layer limit in workload generator

* use TimeSlotAllocation for coala scheduling

* add pydot to requirements.txt

* working 2 layer conv for both standard coala and new extra CO scheduler

* black and broken rebase fixes

---------

+++ stream/hardware/architecture/utils.py
@@ -50,7 +50,7 @@ def intersections(a: list[Any], b: list[Any]):
 
 
 def get_core_capacities(accelerator: "Accelerator", mem_op: MemoryOperand, core_ids: list[int]):
-    core_capacities: dict[str, int] = {}
+    core_capacities: dict[str, float] = {}
     for core_id in core_ids:
         core_name = f"Core {core_id}"
         core = accelerator.get_core(core_id)
